Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/animation-frames/callback-exception-expected.txt

CONSOLE MESSAGE: line 22: Error: requestAnimationFrameException

PASS requestAnimationFrame callback exceptions are reported to error handler 
